Create userclone.ps1
This script allows administrators to copy memberships from one user to another in Azure AD and Exchange Online. The script takes in three parameters: SourceUserAccount, TargetUserAccount, and AdminUPN. The first two parameters are the User Principal Names (UPN) of the source and target users, respectively. The AdminUPN parameter is used to connect to Azure AD and Exchange Online.

The script first connects to Azure AD using the Connect-AzureAD cmdlet with the AdminUPN parameter provided by the user. It then gets the source and target users using the Get-AzureADUser cmdlet with their corresponding UPNs. If both users exist, the script gets all group memberships of the source user using the Get-AzureADUserMembership cmdlet and filters them to only include groups.

Next, the script loops through each group and checks if the target user is not already a member of the group. If the target user is not a member, it adds the target user to the group using the Add-AzureADGroupMember cmdlet.

After completing the Azure AD portion, the script connects to Exchange Online PowerShell using the New-PSSession cmdlet with the AdminUPN parameter provided by the user. It then gets all mail-enabled security groups that the source user is a member of using the Get-DistributionGroup cmdlet.

Finally, the script adds the target user to each security group using the Add-DistributionGroupMember cmdlet. Once all security groups have been processed, the script disconnects from Exchange Online PowerShell using the Remove-PSSession cmdlet.

If either the source or target user is not valid, the script displays a warning message using the Write-Warning cmdlet. Otherwise, it displays a success message for each group and security group that the target user was added to using the Write-Output cmdlet.